tests_automation/requirements.txt
# tests_automation/requirements.txt
pytest
pytest-html
pytest-xdist
requests
selenium
Faker
locust
allure-pytest
python-dotenv
